+++ src/WebServer.h
@@ -5,16 +5,18 @@
 #include <LittleFS.h>
 #include <ArduinoJson.h>
 #include <vector>
+#include "TimeManager.h"
 #include "RaceHistory.h"
 #include "Version.h"
 #include "Configuration.h"
+#include "NetworkManager.h"
 
 // Function pointer type for command handler
 typedef void (*CommandHandler)(const char* command);
 
 class WebServer {
 public:
-    WebServer(TimeManager& timeManager, Configuration& config);
+    WebServer(TimeManager& tm, Configuration& cfg, NetworkManager& nm);
     void begin();
     void handleWebSocketMessage(AsyncWebSocketClient *client, const char *data);
     void notifyStatus(const char* status);
@@ -23,6 +25,7 @@ public:
     void notifyRaceComplete(float lane1, float lane2);
     void sendVersionInfo(AsyncWebSocketClient *client);
     void setCommandHandler(CommandHandler handler);
+    void notifyNetworkStatus();
     
 private:
     AsyncWebServer server;
@@ -32,9 +35,11 @@ private:
     CommandHandler commandHandler;
     RaceHistory raceHistory;  // Will be initialized in constructor
     Configuration& config;
+    NetworkManager& networkManager;
     void onWebSocketEvent(AsyncWebSocket *server, AsyncWebSocketClient *client,
                          AwsEventType type, void *arg, uint8_t *data, size_t len);
     void setupRoutes();
     void broadcastJson(const JsonDocument& doc);
     void sendRaceHistory(AsyncWebSocketClient *client);
+    void sendNetworkInfo(AsyncWebSocketClient *client);
 };